SendPacketValidationContext
