+++ crates/ibc/src/test_utils.rs
@@ -16,7 +16,7 @@ use crate::core::ics03_connection::connection::ConnectionEnd;
 use crate::core::ics03_connection::error::ConnectionError;
 use crate::core::ics04_channel::channel::{ChannelEnd, Counterparty, Order};
 use crate::core::ics04_channel::commitment::PacketCommitment;
-use crate::core::ics04_channel::context::SendPacketReader;
+use crate::core::ics04_channel::context::SendPacketValidationContext;
 use crate::core::ics04_channel::error::{ChannelError, PacketError};
 use crate::core::ics04_channel::handler::ModuleExtras;
 use crate::core::ics04_channel::msgs::acknowledgement::Acknowledgement;
@@ -309,7 +309,7 @@ impl TokenTransferReader for DummyTransferModule {
     }
 }
 
-impl SendPacketReader for DummyTransferModule {
+impl SendPacketValidationContext for DummyTransferModule {
     fn channel_end(&self, chan_end_path: &ChannelEndPath) -> Result<ChannelEnd, ContextError> {
         match self
             .ibc_store
@@ -408,6 +408,26 @@ impl SendPacketReader for DummyTransferModule {
         use sha2::Digest;
 
         sha2::Sha256::digest(value).to_vec()
+    }
+
+    fn compute_packet_commitment(
+        &self,
+        packet_data: &[u8],
+        timeout_height: &crate::core::ics04_channel::timeout::TimeoutHeight,
+        timeout_timestamp: &crate::timestamp::Timestamp,
+    ) -> PacketCommitment {
+        let mut hash_input = timeout_timestamp.nanoseconds().to_be_bytes().to_vec();
+
+        let revision_number = timeout_height.commitment_revision_number().to_be_bytes();
+        hash_input.append(&mut revision_number.to_vec());
+
+        let revision_height = timeout_height.commitment_revision_height().to_be_bytes();
+        hash_input.append(&mut revision_height.to_vec());
+
+        let packet_data_hash = self.hash(packet_data);
+        hash_input.append(&mut packet_data_hash.to_vec());
+
+        self.hash(&hash_input).into()
     }
 }
